+++ libmpeg2/mpeg2.h
@@ -87,7 +87,7 @@ typedef struct mpeg2_gop_s {
 typedef struct mpeg2_picture_s {
     unsigned int temporal_reference;
     unsigned int nb_fields;
-    uint32_t tag, tag2;
+    __int64 tag, tag2;
     uint32_t flags;
     struct {
 	int x, y;
@@ -177,7 +177,7 @@ void mpeg2_reset (mpeg2dec_t * mpeg2dec, int full_reset);
 void mpeg2_skip (mpeg2dec_t * mpeg2dec, int skip);
 void mpeg2_slice_region (mpeg2dec_t * mpeg2dec, int start, int end);
 
-void mpeg2_tag_picture (mpeg2dec_t * mpeg2dec, uint32_t tag, uint32_t tag2);
+void mpeg2_tag_picture (mpeg2dec_t * mpeg2dec, __int64 tag, __int64 tag2);
 
 void mpeg2_init_fbuf (mpeg2_decoder_t * decoder, uint8_t * current_fbuf[3],
 		      uint8_t * forward_fbuf[3], uint8_t * backward_fbuf[3]);
